Move info grid to own file
and minor formatting
